requirements.txt
azure-functions>=1.20.0
requests>=2.31.0
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
azure-mgmt-devcenter>=1.0.0
azure-developer-devcenter>=1.0.0
python-dateutil>=2.8.0
urllib3